move EnrollmentGraph folder into RightPane folder
