+++ src/components/RightPane/EnrollmentGraph/static/quarter_day_ranges.json
